[2025-01-17 JPA STUDY 보완(2)] - file Service 분리 - 비효율적인 코드 개선 - 엔티티 id 타입 수정 - repository 메서드 수정 - Log 수정 - html 태그 오류 수정

+++ PeachBoard/src/main/java/com/board/dto/PostDetailRes.java
@@ -19,7 +19,7 @@ import lombok.Setter;
 @Builder
 public class PostDetailRes {
 
-	private Integer id; // 게시글 아이디
+	private Long id; // 게시글 아이디
 
 	private String writer; // 작성자
